+++ src/Annotation/Meta.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace JSONAPI\Annotation;
+
+/**
+ * Class Meta
+ *
+ * @package JSONAPI\Annotation
+ * @Annotation
+ * @Target({"METHOD", "PROPERTY"})
+ */
+class Meta extends Common
+{
+
+}